ai_backend/pyproject.toml
[project]
name = "ai-backend"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.32.0",
  "google-genai>=1.29.0",
  "python-dotenv>=1.0.0",
  "websockets>=13.0",
  "litellm>=1.82.6",
]
